refactor: rename ValidationError2 to SpecError
