+++ src/plccng/spec/SpecError.py
